+++ language/german.php
@@ -378,6 +378,13 @@ $LANG_PP = array (
 'dscp_purge_carts' => 'Delete all customer shopping carts that have not been finalized as orders.<br /><br>This will impact the guest experience and should not normally be necessary.</b>',
 'include' => 'Include',
 'exclude' => 'Exclude',
+'dscp_change_currency' => 'Change prices for products, attributes and shopping cart items to a new currency.<br /><b>This will adjust product prices. You should review all prices for accuracy following this operation.</b>',
+'change_currency' => 'Change Currency',
+'old_currency' => 'Old Currency Setting',
+'convert_to' => 'Convert To',
+'q_change_currency' => 'Are you sure you want to change prices and other amounts based on the new currency?',
+'cur_changed' => 'Prices changed from %1$s to %2%s',
+'no_cur_change' => 'No currency conversion was done.',
 );
 if (isset($_PP_CONF['ena_ratings']) && $_PP_CONF['ena_ratings']) {
     $LANG_PP['list_sort_options']['top_rated'] = 'Top Rated';